Move ui components to src/app/_components directory
