+++ resumex/src/components/login-form.tsx
@@ -1,15 +1,13 @@
-import { cn } from "~/lib/utils"
-import { Button } from "~/components/ui/button"
+import { Button } from "~/app/_components/button";
 import {
   Card,
   CardContent,
   CardDescription,
   CardHeader,
   CardTitle,
-} from "~/components/ui/card"
-import { Input } from "~/components/ui/input"
-import { Label } from "~/components/ui/label"
-import Link from "next/link"
+} from "~/app/_components/card";
+import { Input } from "~/app/_components/input";
+import { Label } from "~/app/_components/label";
 
 export function LoginForm({
   className,
